// setup/src/types.ts
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the Elastic License 2.0;
 * you may not use this file except in compliance with the Elastic License 2.0.
 */

export type ConnectionType = 'local' | 'cloud-hosted' | 'cloud-serverless';

export interface SetupConfig {
  connectionType: ConnectionType;
  cloudId?: string;
  esNode?: string;
  apiKey?: string;
  username?: string;
  password?: string;
  kibanaUrl?: string;
  unsafeSsl: boolean;
}